statistics.py done
